requirements.txt
asgiref==3.8.1
black==23.12.1
click==8.1.8
django==4.2.11
djangorestframework==3.15.2
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.6
sqlparse==0.5.3
typing-extensions==4.13.2
psycopg==3.2.9
psycopg2-binary==2.9.10
sqlparse==0.5.3
whitenoise==6.7.0